+++ scripts/chicago_isac_scraper.py
@@ -15,6 +15,7 @@ request_urls = [
     'https://isac.uchicago.edu/research/publications/chicago-assyrian-dictionary',
     'https://isac.uchicago.edu/research/publications/chicago-hittite-dictionary',
     'https://isac.uchicago.edu/research/publications/chicago-hittite-dictionary-supplements-chds',
+    'https://isac.uchicago.edu/research/publications/oriental-institute-seminars-ois',
     'https://isac.uchicago.edu/research/publications/late-antique-and-medieval-islamic-near-east-lamine',
     'https://isac.uchicago.edu/research/publications/materials-assyrian-dictionary',
     'https://isac.uchicago.edu/research/publications/materials-and-studies-kassite-history-mskh',
@@ -23,7 +24,6 @@ request_urls = [
     'https://isac.uchicago.edu/research/publications/oriental-institute-digital-archives-oida',
     'https://isac.uchicago.edu/research/oimp',
     'https://isac.uchicago.edu/research/publications/oriental-institute-publications-oip',
-    'https://isac.uchicago.edu/research/publications/oriental-institute-seminars-ois',
     'https://isac.uchicago.edu/research/publications/studies-ancient-oriental-civilization-saoc',
     'https://isac.uchicago.edu/research/publications/lost-egypt',
     'https://isac.uchicago.edu/research/publications/miscellaneous-publications'
@@ -32,6 +32,14 @@ request_urls = [
 finalMetaData = []
 
 OIC_ROOT = 'https://isac.uchicago.edu/'
+
+isbnPattern = re.compile(r'ISBN\-10 |ISBN\-10|ISBN\-13 |ISBN\-13|ISBN |ISBN|\:|\-|\
+                            \(hardback\) | \(hardcover\) |\(eBook\) ')
+charPattern = re.compile(r'\n|\t|\\|\\:|\u00a0')
+titlePattern = re.compile(r'AS [\d]+. |CHDS [\d]+. |ISAC [\d]+. | OIS [\d]+. |LAMINE [\d]+. |\
+                            MAD [\d]+. |MSKH [\d] |NE [\d]+. |OIC [\d]+. |OIDA [\d]+. |OIMP [\d]+. |\
+                            OIP [\d]+. |SAOC [/d]+. ')
+authorPattern = re.compile(r'ed\.|eds\.|([\d]+)\.| and |([\d]+)|Originally published in ([\d]+)\.|\n')
 
 def main(): 
 
@@ -48,22 +56,28 @@ def main():
             url == 'https://isac.uchicago.edu/research/publications/chicago-hittite-dictionary' or \
             url == 'https://isac.uchicago.edu/research/publications/lost-egypt':
 
-            parsePubWithSamePageDownload(elem)
+            parsePubWithSamePageDownload(elem, url)
 
         else: 
             parsePubWithDownloadRedirect(elem, url)
 
-    finalMetaDataJSON = cleaningText(finalMetaData)
+    finalMetaDataJSON = cleanText(finalMetaData)
     print(len(finalMetaDataJSON))
 
-    with open("data_file.json", "w", encoding='utf-8') as write_file:
+    with open("chicagoISAC_metadata.json", "w", encoding='utf-8') as write_file:
         json.dump(finalMetaDataJSON, write_file, ensure_ascii = False, indent = 6)
 
-def parsePubWithSamePageDownload(elem):
+def parsePubWithSamePageDownload(elem, url):
 
     '''Parsing publications with metadata present on same page as download link'''
 
     soup = BeautifulSoup(elem.text, 'lxml')   #Parsing the catalog's webpage using BeautifulSoup
+
+    titleContainer = soup.find(id='content-top')
+
+    spanContainer = titleContainer.find('span')
+
+    titleEntry = spanContainer.text
         
     catContainer = soup.find(class_='catalog')
 
@@ -74,20 +88,55 @@ def parsePubWithSamePageDownload(elem):
     for pub in pubEntry:
         pub_tds = pub.find_all('td')
         if pub_tds[5].find('a', class_="publication ss-standard ss-download btn"):
-            finalMetaData.append({
-            'Volume': pub_tds[0].text,
-            'publicationInfo': pub_tds[2].text,
-            'isbn': pub_tds[3].text,
-            'Extent': pub_tds[1].text,
-            'url': [pub_tds[5].find('a', class_="publication ss-standard ss-download btn").get('href', None)]
-        })
-            print(finalMetaData[1]['isbn'])
+            
+            urlStr = pub_tds[5].find('a', class_="publication ss-standard ss-download btn").get('href', None)
+            
+            if url == 'https://isac.uchicago.edu/research/publications/lost-egypt':
+                finalMetaData.append({
+                    'title': titleEntry,
+                    'authors': ['The Epigraphic Survey of the Oriental Institute of the University of Chicago'],
+                    'series': pub_tds[0].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(pub_tds[2].text)['publisher'],
+                    'publicationDate': pubParsing(pub_tds[2].text)['date'],
+                    'isbn': pub_tds[3].text,
+                    'extent': pub_tds[1].text,
+                    'url': [f'{OIC_ROOT}{urlStr}']
+                })
+            elif url == 'https://isac.uchicago.edu/research/publications/chicago-assyrian-dictionary':
+                finalMetaData.append({
+                    'title': titleEntry,
+                    'authors': ['Robert D. Biggs, John A. Brinkman, Miguel Civil, Walter Farber, Erica Reiner, Martha T. Roth, Matthew W. Stolper'],
+                    'series': pub_tds[0].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(pub_tds[2].text)['publisher'],
+                    'publicationDate': pubParsing(pub_tds[2].text)['date'],
+                    'isbn': pub_tds[3].text,
+                    'extent': pub_tds[1].text,
+                    'url': [f'{OIC_ROOT}{urlStr}']
+                })
+            else:
+                finalMetaData.append({
+                    'title': titleEntry,
+                    'authors': ['Petra M. Goedegebuure, Hans G. Güterbock, Harry A. Hoffner, Theo P. J. van den Hout'],
+                    'series': pub_tds[0].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(pub_tds[2].text)['publisher'],
+                    'publicationDate': pubParsing(pub_tds[2].text)['date'],
+                    'isbn': pub_tds[3].text,
+                    'extent': pub_tds[1].text,
+                    'url': [f'{OIC_ROOT}{urlStr}']
+                })
         else:
             finalMetaData.append({
-            'Volume': pub_tds[0].text,
-            'publicationInfo': pub_tds[2].text,
+            'title': titleEntry,
+            'authors': ['Petra M. Goedegebuure, Hans G. Güterbock, Harry A. Hoffner, and Theo P. J. van den Hout'],
+            'series': pub_tds[0].text,
+            'publisherLocation': 'Chicago',
+            'publisher': pubParsing(pub_tds[2].text)['publisher'],
+            'publicationDate': pubParsing(pub_tds[2].text)['date'],
             'isbn': pub_tds[3].text,
-            'Extent': pub_tds[1].text,
+            'extent': pub_tds[1].text,
             'url': []
         })
 
@@ -114,13 +163,16 @@ def parsePubWithDownloadRedirect(elem, url):
         if pub.find('a', class_= "publication ss-standard ss-download btn"):
             downloadLinksList = pub.find_all('a', class_= "publication ss-standard ss-download btn")
             for link in downloadLinksList:
-                downloadLinks.append(link.get('href', None))
+                urlStr = link.get('href', None)
+                downloadLinks.append(f'{OIC_ROOT}{urlStr}')
             
     
         pubLinks = pub.find_all('a')
         
         #Link should only be the link to the publication webpage
         detailLink = pubLinks[0].get('href', None)
+        detailTitle = pubLinks[0].text
+        detailAuthor = pubLinks[0].next_sibling.text
 
         detailURL = detailLinkUpdate(detailLink)
         print(detailURL)
@@ -136,7 +188,7 @@ def parsePubWithDownloadRedirect(elem, url):
             detailURL == 'https://isac.uchicago.edu//node/3390' or\
             detailURL == 'https://isac.uchicago.edu//node/3380' or\
             detailURL == 'https://isac.uchicago.edu//node/3338':
-            parsePubMISC(detailURL, pubContainer, downloadLinks)
+            parsePubMISC(detailURL, detailTitle, detailAuthor, pubContainer, downloadLinks)
             continue
 
         # Metadata list is the last unordered list on the webpage
@@ -148,8 +200,12 @@ def parsePubWithDownloadRedirect(elem, url):
             if len(metadata) > 5:
 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[1].text,
-                    'publicationInfo': metadata[2].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[2].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[2].text)['date'],
                     'isbn': metadata[3].text,
                     'extent': metadata[4].text,
                     'url': downloadLinks
@@ -157,8 +213,12 @@ def parsePubWithDownloadRedirect(elem, url):
             else:
 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[0].text,
-                    'publicationInfo': metadata[1].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[1].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[1].text)['date'],
                     'isbn': metadata[2].text,
                     'extent': metadata[3].text,
                     'url': downloadLinks
@@ -168,8 +228,12 @@ def parsePubWithDownloadRedirect(elem, url):
             if len(metadata) == 5 and 'Seminars' in metadata[0].text:
 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[0].text,
-                    'publicationInfo': metadata[1].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[1].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[1].text)['date'],
                     'isbn': metadata[2].text,
                     'extent': metadata[3].text,
                     'url': downloadLinks
@@ -177,7 +241,12 @@ def parsePubWithDownloadRedirect(elem, url):
             elif len(metadata) == 5 and 'Seminars' not in metadata[0].text:
 
                 finalMetaData.append({
-                    'publicationInfo': metadata[0].text,
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
+                    'series': '',
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[0].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[0].text)['date'],
                     'isbn': metadata[1].text,
                     'extent': metadata[2].text,
                     'url': downloadLinks
@@ -185,8 +254,12 @@ def parsePubWithDownloadRedirect(elem, url):
             elif len(metadata) == 6 and 'Seminars' in metadata[3].text:
 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[3].text,
-                    'publicationInfo': metadata[0].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[0].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[0].text)['date'],
                     'isbn': metadata[1].text,
                     'extent': metadata[2].text,
                     'url': downloadLinks
@@ -194,8 +267,12 @@ def parsePubWithDownloadRedirect(elem, url):
             else:
                 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[0].text,
-                    'publicationInfo': metadata[1].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[1].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[1].text)['date'],
                     'isbn': metadata[2].text,
                     'extent': metadata[3].text,
                     'url': downloadLinks
@@ -206,7 +283,13 @@ def parsePubWithDownloadRedirect(elem, url):
             and (len(metadata) == 2 or len(metadata) == 3):
             
             finalMetaData.append({
-                'publicationInfo': metadata[0].text,
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
+                'series': '',
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[0].text)['publisher'],
+                'publicationDate': pubParsing(metadata[0].text)['date'],
+                'isbn': '',
                 'extent': metadata[1].text,
                 'url': downloadLinks
             }) 
@@ -215,17 +298,29 @@ def parsePubWithDownloadRedirect(elem, url):
             and len(metadata) == 2:
 
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
+                'isbn': '',
+                'extent': '',
                 'url': downloadLinks
             }) 
         
         elif url == 'https://isac.uchicago.edu/research/publications/oriental-institute-communications-oic'\
             and detailLink == '/node/3305':
                 
-                finalMetaData.append({
+            finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
+                'isbn': '',
+                'extent': '',
                 'url': downloadLinks
             }) 
 
@@ -233,8 +328,13 @@ def parsePubWithDownloadRedirect(elem, url):
             and detailLink == '/node/3295':
                 
                 finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
+                'isbn': '',
                 'extent': metadata[2].text,
                 'url': downloadLinks
             })
@@ -247,8 +347,13 @@ def parsePubWithDownloadRedirect(elem, url):
             or detailLink == '/node/3254'):
                 
                 finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
+                'isbn': '',
                 'extent': metadata[2].text,
                 'url': downloadLinks
             })
@@ -258,9 +363,13 @@ def parsePubWithDownloadRedirect(elem, url):
             and 'ISBN' in metadata[2].text and 'ISBN' in metadata[3].text:
             
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[0].text,
-                    'publicationInfo': metadata[1].text,
-                    'isbn': [metadata[2].text,metadata[3].text],
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[1].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[1].text)['date'],
+                    'isbn': f'{metadata[2].text}, {metadata[3].text}',
                     'extent': metadata[4].text,
                     'url': downloadLinks
                 })
@@ -268,8 +377,12 @@ def parsePubWithDownloadRedirect(elem, url):
         elif 'ISBN' in metadata[2].text:
 
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
                 'isbn': metadata[2].text,
                 'extent': metadata[3].text,
                 'url': downloadLinks
@@ -278,8 +391,12 @@ def parsePubWithDownloadRedirect(elem, url):
         elif len(metadata) > 3 and 'ISBN' in metadata[3].text:
 
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
                 'isbn': metadata[3].text,
                 'extent': metadata[2].text,
                 'url': downloadLinks
@@ -289,8 +406,13 @@ def parsePubWithDownloadRedirect(elem, url):
             len(metadata) == 5:
                 
                 finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': [metadata[1].text, metadata[2].text],
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[2].text)['publisher'],
+                'publicationDate': pubParsing(metadata[2].text)['date'],
+                'isbn': '',
                 'extent': metadata[3].text,
                 'url': downloadLinks
                 }) 
@@ -299,8 +421,13 @@ def parsePubWithDownloadRedirect(elem, url):
             len(metadata) > 7:
                 
                 finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': [metadata[1].text, metadata[2].text, metadata[3].text],
+                'publisherLocation': 'Chicago',
+                #Publisher split in 3 list blocks: metadata[1] metadata[2] metadata[3]
+                'publisher': pubParsing(metadata[2].text)['publisher'],
+                'publicationDate': pubParsing(metadata[2].text)['date'],
                 'isbn': metadata[4].text,
                 'extent': metadata[5].text,
                 'url': downloadLinks
@@ -309,8 +436,13 @@ def parsePubWithDownloadRedirect(elem, url):
         else:
 
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
+                'isbn': '',
                 'extent': metadata[2].text,
                 'url': downloadLinks
             }) 
@@ -320,13 +452,17 @@ def parsePubSAOC(pub, url, downloadLinks):
         
     #Link should only be the link to the publication webpage
     detailLink = pubLinks[0].get('href', None)
+    detailTitle = pubLinks[0].text
+    detailAuthor = pubLinks[0].next_sibling.text
 
+    #When download Link and Link to publication not in same List, add download link to previous metadata object
     if pubLinks[0].has_attr('class') == True:
         if 'url' not in finalMetaData[-1]:
 
             downloadLinksList = pub.find_all('a', class_= "publication ss-standard ss-download btn")
             for link in downloadLinksList:
-                downloadLinks.append(link.get('href', None))
+                urlString = link.get('href', None)
+                downloadLinks.append(f'{OIC_ROOT}{urlString}')
             finalMetaData[-1]['url'] = downloadLinks
 
     elif pub.find('a', class_= "publication ss-standard ss-download btn") != None and pubLinks[0].has_attr('class') == False:
@@ -358,23 +494,32 @@ def parsePubSAOC(pub, url, downloadLinks):
 
         downloadLinksList = pub.find_all('a', class_= "publication ss-standard ss-download btn")
         for link in downloadLinksList:
-            downloadLinks.append(link.get('href', None))
+            urlString = link.get('href', None)
+            downloadLinks.append(f'{OIC_ROOT}{urlString}')
 
         if 'ISBN' in metadata[2].text:
             if detailURL == 'https://isac.uchicago.edu//node/3112':
 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[0].text,
-                    'publicationInfo': metadata[1].text,
-                    'isbn': [metadata[2].text,metadata[3].text],
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[1].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[1].text)['date'],
+                    'isbn': f'{metadata[2].text}, {metadata[3].text}',
                     'extent': metadata[4].text,
                     'url': downloadLinks
                 }) 
             else:
 
                 finalMetaData.append({
+                    'title': detailTitle,
+                    'authors': cleanAuthorMetadata(detailAuthor),
                     'series': metadata[0].text,
-                    'publicationInfo': metadata[1].text,
+                    'publisherLocation': 'Chicago',
+                    'publisher': pubParsing(metadata[1].text)['publisher'],
+                    'publicationDate': pubParsing(metadata[1].text)['date'],
                     'isbn': metadata[2].text,
                     'extent': metadata[3].text,
                     'url': downloadLinks
@@ -382,8 +527,13 @@ def parsePubSAOC(pub, url, downloadLinks):
         else:
 
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
+                'isbn': '',
                 'extent': metadata[2].text,
                 'url': downloadLinks
             })
@@ -409,18 +559,27 @@ def parsePubSAOC(pub, url, downloadLinks):
 
             cleanISBN = metadata[2].text.replace("\u00a0", "")
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
-                'isbn': cleanISBN,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
                 'extent': metadata[3].text,
+                'isbn': cleanISBN,
             }) 
 
         else:
 
             finalMetaData.append({
+                'title': detailTitle,
+                'authors': cleanAuthorMetadata(detailAuthor),
                 'series': metadata[0].text,
-                'publicationInfo': metadata[1].text,
+                'publisherLocation': 'Chicago',
+                'publisher': pubParsing(metadata[1].text)['publisher'],
+                'publicationDate': pubParsing(metadata[1].text)['date'],
                 'extent': metadata[2].text,
+                'isbn': '',
             })
 
 def detailLinkUpdate(detailLink):
@@ -430,39 +589,82 @@ def detailLinkUpdate(detailLink):
     if detailLink == 'https://www.isdistribution.com/Refer.aspx?isbn=9781885923653':
         detailURL = 'https://isac.uchicago.edu/research/publications/oimp/oimp-31-ancient-israel-highlights-collections-oriental-institute'
         return detailURL
+    elif detailLink == 'https://www.isdistribution.com/Refer.aspx?isbn=9781614910855':
+        detailURL = 'https://isac.uchicago.edu/research/publications/isacs/isacs14'
+        return detailURL
     else:
         detailURL = f'{OIC_ROOT}{detailLink}'
         return detailURL
     
-def cleaningText(metaData):
+def cleanText(metaData):
     for i, dataDict in enumerate(metaData):
         for key, value in dataDict.items():
             if type(value) == str:
-                metaData[i][key] = re.sub(r'[\n\t\\\"]', '', metaData[i][key])
-                metaData[i][key] = re.sub(r'[\u00a0]', ' ', metaData[i][key])
+                if key == 'isbn':
+                    metaData[i][key] = re.sub(isbnPattern, '', metaData[i][key])
+                metaData[i][key] = re.sub(charPattern, '', metaData[i][key])
+                metaData[i][key] = re.sub(titlePattern, '', metaData[i][key])
+                metaData[i][key] = metaData[i][key].strip()
+    
     return metaData
 
+def cleanAuthorMetadata(authors):
+    cleanAuthors = re.sub(authorPattern, '', authors)
+    cleanAuthorsList = cleanAuthors.strip().split(', ')
+    return cleanAuthorsList
+
+def pubParsing(pubInfo):
+    pubDictionary = {'publisher': [], 'date': ''}
+    pubSplitInfo = re.split(r"([\d]+)", pubInfo)
+    pubParsedList = [info.strip() for info in pubSplitInfo]
+    pubParsedInfo = list(filter(None, pubParsedList))
+    print(pubParsedInfo)
     
-def parsePubMISC(detailURL, pubContainer, downloadLinks):
+    if pubParsedInfo[0] == '-' or pubParsedInfo == '':
+        return pubDictionary
+    elif pubParsedInfo[0].isdigit():
+        pubDictionary['date'] = pubParsedInfo[0]
+    elif pubParsedInfo[0].isdigit() == False and len(pubParsedInfo) == 1:
+        pubParsedInfoComma = pubParsedInfo[0].replace(',', '')
+        pubParsedInfoLocation = pubParsedInfoComma.replace('Chicago: ', '')
+        pubDictionary['publisher'].append(pubParsedInfoLocation)
+        pubDictionary['publisher']
+    else:
+        pubParsedInfoComma = pubParsedInfo[0].replace(',', '')
+        pubParsedInfoLocation = pubParsedInfoComma.replace('Chicago: ', '')
+        pubDictionary['publisher'].append(pubParsedInfoLocation)
+        pubDictionary['date'] = pubParsedInfo[1]
+    return pubDictionary
+
+    
+def parsePubMISC(detailURL, detailTitle, detailAuthor, pubContainer, downloadLinks):
     if detailURL == 'https://isac.uchicago.edu//node/3788':
         dataList = pubContainer.find_all('p')[-1]
         print(dataList)
 
         line_break =  dataList.find('br')       # loop through line break tags
         line_break.replaceWith(' ')       # replace br tags with delimiter
-        strings = dataList.get_text().split('\n')
+        metadata = dataList.get_text().split('\n')
         finalMetaData.append({
-        'series': strings[0],
-        'publicationInfo': strings[1],
-        'isbn': strings[4],
-        'extent': strings[1],
-        'url': downloadLinks
+            'title': detailTitle,
+            'authors': cleanAuthorMetadata(detailAuthor),
+            'series': metadata[0],
+            'publisherLocation': 'Chicago',
+            'publisher': pubParsing(metadata[1])['publisher'],
+            'publicationDate': pubParsing(metadata[1])['date'],
+            'isbn': metadata[4],
+            'extent': metadata[1],
+            'url': downloadLinks
         }) 
     
     elif detailURL == 'https://isac.uchicago.edu//node/3990':
         finalMetaData.append({
+            'title': detailTitle,
+            'authors': cleanAuthorMetadata(detailAuthor),
             'series': 'Reliefs and Inscriptions from the Tomb of Per-Haps: An Oriental Institute Holiday Card',
-            'publicationInfo': '',
+            'publisherLocation': 'Chicago',
+            'publisher': [],
+            'publicationDate': '',
             'isbn': '',
             'extent': '',
             'url': downloadLinks
@@ -477,8 +679,12 @@ def parsePubMISC(detailURL, pubContainer, downloadLinks):
 
 
         finalMetaData.append({
-            'publicationInfo': [metadata[0].text, metadata2[0].text],
-            'extent': [metadata[1].text, metadata2[1].text]
+            'title': detailTitle,
+            'authors': cleanAuthorMetadata(detailAuthor),
+            'publisherLocation': 'Chicago',
+            'publisher': ['Joint Publication with the Cotsen Institute of Archaeology, University of California, Los Angeles'],
+            'publicationDate': '2007',
+            'extent': f'{metadata[1].text}, {metadata2[1].text}'
         })
 
     elif detailURL == 'https://isac.uchicago.edu//node/3380':
@@ -488,8 +694,12 @@ def parsePubMISC(detailURL, pubContainer, downloadLinks):
         metadata = dataList.find_all('li')
 
         finalMetaData.append({
+            'title': detailTitle,
+            'authors': cleanAuthorMetadata(detailAuthor),
             'series': metadata[0].text,
-            'publicationInfo': metadata[2].text,
+            'publisherLocation': 'Chicago',
+            'publisher': pubParsing(metadata[2].text)['publisher'],
+            'publicationDate': pubParsing(metadata[2].text)['date'],
             'isbn': metadata[1].text,
             'extent': metadata[3].text,
             'url': downloadLinks
@@ -502,8 +712,12 @@ def parsePubMISC(detailURL, pubContainer, downloadLinks):
         metadata = dataList.find_all('li')
 
         finalMetaData.append({
+            'title': detailTitle,
+            'authors': cleanAuthorMetadata(detailAuthor),
             'series': metadata[0].text,
-            'publicationInfo': [metadata[1].text, metadata[2].text],
+            'publisherLocation': 'Chicago',
+            'publisher': pubParsing(metadata[2].text)['publisher'],
+            'publicationDate': pubParsing(metadata[2].text)['date'],
             'extent': metadata[3].text,
             'url': downloadLinks
             })